requirements.txt
autopep8==1.7.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
flake8==5.0.4
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.4
greenlet==1.1.3.post0
httplib2==0.20.4
idna==3.4
mccabe==0.7.0
openpyxl==3.0.10
protobuf==4.21.8
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.42
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.12
